Moved extract_isolate_source() to utils
